+++ device-management-application/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Conanfile & Conan CI
 - [SW-13480]: Adding memshire frequency log print
+- et-power-top: Printing compute utilization stats.
 ### Changed
 ### Deprecated
 ### Removed